src/requirements.txt
Django==3.0.4
psycopg2==2.8.5
gunicorn==20.0.4
requests==2.23.0
Pillow==7.1.1
PyJWT==1.7.1
django-cors-headers==3.2.1
django-crispy-forms==1.9
djangorestframework==3.11.0
six==1.14
